Small cleanups in db.js
refStaticAsset specified an incorrect query filter when attempting to increment
the refcount of the asset in question.  This was discovered by me playing
around with eslint rules, which detected the use of an undefined "hash" in that
function.

A couple assignments were being made to the global scope when they should
probably be local scope.

I folded a pair of if (Meteor.isServer) blocks together.