First pass at a bpf_asm based seccomp filter.
Untested, not yet integrated into the supervisor. Note that the
filter is not equivalent to the one we build with libseccomp.
Instead, it is a whitelist that we will no doubt end up needing
to expand. It is based on an earlier attempt of mine to do
a whitelist version with libseccomp.