Some fixes to setsockopt() filtering.
I started to expand the list of what we allow, and then noticed that we
had jeq in some places where we wanted jne. I've left the stubs for
expanded lists in place for now, since we'll probably need to do that
later, but this gets us a bit further.