Clarify/prescribe the behavior of Assignable.Getter.subscribe()
We don't seem to already have any Getters in our codebase, so this
trivially remains consistent with existing behavior.

The semantics now described are easier to work with, since usually you
want the initial value *and* updates, so this obviates needing to also
call .get() and worry about race conditions, etc.